requirements.txt
aniso8601==9.0.1
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
mariadb==1.0.6
MarkupSafe==2.0.1
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
PyYAML==5.4.1
safrs==2.11.4
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
